+++ src/main.rs
@@ -144,6 +144,8 @@ async fn main() {
         .route("/policy/{name}/rules/{id}/toggle",    post(routes::rules::post_rule_toggle))
         .route("/policy/{name}/rules/{id}/delete",    post(routes::rules::post_rule_delete))
         .route("/rules",                 get(routes::rules::get_all_rules))
+        .route("/rules/new",             get(routes::rules::get_custom_rule_new))
+        .route("/rules/create",          post(routes::rules::post_custom_rule_create))
         .route("/rules/{id}/edit",       get(routes::rules::get_rule_edit_global))
         .route("/rules/{id}/update",     post(routes::rules::post_rule_update_global))
         .route("/rules/{id}/toggle",     post(routes::rules::post_rule_toggle_global))